requirements.txt
flask
flask-cors
openai
pinecone
python-dotenv
gunicorn